+++ frontend/src/pages/AdminSongsPage.tsx
@@ -1,88 +1,11 @@
-import {
-  useCallback,
-  useEffect,
-  useMemo,
-  useRef,
-  useState,
-  type ChangeEvent,
-  type FormEvent,
-} from "react";
+import { useState, type FormEvent } from "react";
 import { ConfirmDialog } from "../components/ConfirmDialog";
-import { Skeleton } from "../components/Skeleton";
+import { SongEditForm } from "../components/SongEditForm";
+import { SongTable } from "../components/SongTable";
 import { useToast } from "../context/useToast";
-import {
-  ApiError,
-  bulkImportSongs,
-  createSong,
-  deleteSong,
-  listGenres,
-  listSongs,
-  updateSong,
-} from "../lib/api";
+import { useAdminSongs } from "../hooks/useAdminSongs";
 import { clearAdminPassword, getAdminPassword, setAdminPassword } from "../lib/adminPassword";
-import type { Genre, Song, SongWritePayload } from "../lib/types";
 import styles from "./AdminSongsPage.module.css";
-
-const PER_PAGE = 50;
-const YT_ID_PATTERN = /^[A-Za-z0-9_-]{11}$/;
-const SEARCH_DEBOUNCE_MS = 250;
-
-type Mode = { kind: "list" } | { kind: "create" } | { kind: "edit"; song: Song };
-
-interface FormState {
-  title: string;
-  artist: string;
-  youtube_id: string;
-  start_time: string;
-  release_year: string;
-  genre_ids: Set<string>;
-}
-
-const EMPTY_FORM: FormState = {
-  title: "",
-  artist: "",
-  youtube_id: "",
-  start_time: "0",
-  release_year: "",
-  genre_ids: new Set(),
-};
-
-function songToForm(song: Song, genreIds: string[]): FormState {
-  return {
-    title: song.title,
-    artist: song.artist,
-    youtube_id: song.youtube_id,
-    start_time: String(song.start_time),
-    release_year: song.release_year != null ? String(song.release_year) : "",
-    genre_ids: new Set(genreIds),
-  };
-}
-
-function formToPayload(form: FormState): SongWritePayload {
-  return {
-    title: form.title.trim(),
-    artist: form.artist.trim(),
-    youtube_id: form.youtube_id.trim(),
-    start_time: Number(form.start_time),
-    release_year: form.release_year.trim() === "" ? null : Number(form.release_year),
-    genre_ids: Array.from(form.genre_ids),
-  };
-}
-
-function formIsValid(form: FormState): boolean {
-  if (form.title.trim().length === 0 || form.title.trim().length > 200) return false;
-  if (form.artist.trim().length === 0 || form.artist.trim().length > 200) return false;
-  if (!YT_ID_PATTERN.test(form.youtube_id.trim())) return false;
-  const start = Number(form.start_time);
-  if (!Number.isInteger(start) || start < 0) return false;
-  const year = form.release_year.trim();
-  if (year !== "") {
-    const y = Number(year);
-    if (!Number.isInteger(y) || y < 1900 || y > 2100) return false;
-  }
-  if (form.genre_ids.size === 0) return false;
-  return true;
-}
 
 export function AdminSongsPage() {
   const { toast } = useToast();
@@ -160,143 +83,29 @@ interface SongsConsoleProps {
 }
 
 function SongsConsole({ pw, onAuthFail, onSignOut }: SongsConsoleProps) {
-  const { toast } = useToast();
-  const [songs, setSongs] = useState<Song[]>([]);
-  const [total, setTotal] = useState(0);
-  const [page, setPage] = useState(1);
-  const [searchInput, setSearchInput] = useState("");
-  const [search, setSearch] = useState("");
-  const [genreFilter, setGenreFilter] = useState("");
-  const [genres, setGenres] = useState<Genre[]>([]);
-  const [mode, setMode] = useState<Mode>({ kind: "list" });
-  const [deleteId, setDeleteId] = useState<string | null>(null);
-  const [loading, setLoading] = useState(true);
-  const [busy, setBusy] = useState(false);
-  const fileInputRef = useRef<HTMLInputElement | null>(null);
-  const deleteInFlightRef = useRef(false);
-
-  const reportError = useCallback(
-    (err: unknown, fallback: string) => {
-      if (err instanceof ApiError && err.status === 401) {
-        onAuthFail();
-        return;
-      }
-      toast(err instanceof Error ? err.message : fallback, { variant: "error" });
-    },
-    [onAuthFail, toast],
-  );
-
-  // Load genres once.
-  useEffect(() => {
-    let cancelled = false;
-    void (async () => {
-      try {
-        const list = await listGenres();
-        if (!cancelled) setGenres(list);
-      } catch (err) {
-        if (!cancelled) reportError(err, "Failed to load genres");
-      }
-    })();
-    return () => {
-      cancelled = true;
-    };
-  }, [reportError]);
-
-  // Debounce search input → search.
-  useEffect(() => {
-    const handle = window.setTimeout(() => {
-      setSearch(searchInput);
-      setPage(1);
-    }, SEARCH_DEBOUNCE_MS);
-    return () => window.clearTimeout(handle);
-  }, [searchInput]);
-
-  // Refetch when filter inputs change.
-  const fetchSongs = useCallback(async () => {
-    setLoading(true);
-    try {
-      const res = await listSongs(
-        {
-          page,
-          per_page: PER_PAGE,
-          search: search || undefined,
-          genre: genreFilter || undefined,
-        },
-        pw,
-      );
-      setSongs(res.items);
-      setTotal(res.total);
-    } catch (err) {
-      reportError(err, "Failed to load songs");
-    } finally {
-      setLoading(false);
-    }
-  }, [genreFilter, page, pw, reportError, search]);
-
-  useEffect(() => {
-    void fetchSongs();
-  }, [fetchSongs]);
-
-  const totalPages = Math.max(1, Math.ceil(total / PER_PAGE));
-
-  const songToFilterEdit = useCallback((song: Song) => {
-    setMode({ kind: "edit", song });
-  }, []);
-
-  async function handleSubmitForm(payload: SongWritePayload) {
-    setBusy(true);
-    try {
-      if (mode.kind === "create") {
-        await createSong(payload, pw);
-        toast("Song created", { variant: "success" });
-      } else if (mode.kind === "edit") {
-        await updateSong(mode.song.id, payload, pw);
-        toast("Song updated", { variant: "success" });
-      }
-      setMode({ kind: "list" });
-      await fetchSongs();
-    } catch (err) {
-      reportError(err, "Save failed");
-    } finally {
-      setBusy(false);
-    }
-  }
-
-  async function handleConfirmDelete() {
-    if (deleteId === null || deleteInFlightRef.current) return;
-    deleteInFlightRef.current = true;
-    setBusy(true);
-    try {
-      await deleteSong(deleteId, pw);
-      toast("Song deleted", { variant: "success" });
-      setDeleteId(null);
-      await fetchSongs();
-    } catch (err) {
-      reportError(err, "Delete failed");
-    } finally {
-      setBusy(false);
-      deleteInFlightRef.current = false;
-    }
-  }
-
-  async function handleFile(e: ChangeEvent<HTMLInputElement>) {
-    const file = e.target.files?.[0];
-    if (!file) return;
-    setBusy(true);
-    try {
-      const summary = await bulkImportSongs(file, pw);
-      toast(
-        `Imported ${summary.inserted} new + ${summary.updated} updated (${summary.total} total)`,
-        { variant: "success" },
-      );
-      await fetchSongs();
-    } catch (err) {
-      reportError(err, "Import failed");
-    } finally {
-      setBusy(false);
-      if (fileInputRef.current) fileInputRef.current.value = "";
-    }
-  }
+  const {
+    songs,
+    total,
+    totalPages,
+    page,
+    setPage,
+    searchInput,
+    setSearchInput,
+    genreFilter,
+    setGenreFilter,
+    genres,
+    mode,
+    setMode,
+    deleteId,
+    setDeleteId,
+    loading,
+    busy,
+    fileInputRef,
+    startEdit,
+    handleSubmitForm,
+    handleConfirmDelete,
+    handleFile,
+  } = useAdminSongs({ pw, onAuthFail });
 
   return (
     <main className={styles.shell}>
@@ -317,10 +126,7 @@ function SongsConsole({ pw, onAuthFail, onSignOut }: SongsConsoleProps) {
         />
         <select
           value={genreFilter}
-          onChange={(e) => {
-            setGenreFilter(e.target.value);
-            setPage(1);
-          }}
+          onChange={(e) => setGenreFilter(e.target.value)}
           aria-label="Filter by genre"
         >
           <option value="">All genres</option>
@@ -354,17 +160,10 @@ function SongsConsole({ pw, onAuthFail, onSignOut }: SongsConsoleProps) {
       </div>
 
       {mode.kind !== "list" ? (
-        <SongForm
+        <SongEditForm
           key={mode.kind === "edit" ? mode.song.id : "create"}
           genres={genres}
-          initial={
-            mode.kind === "edit"
-              ? songToForm(
-                  mode.song,
-                  (mode.song.genres ?? []).map((g) => g.id),
-                )
-              : EMPTY_FORM
-          }
+          song={mode.kind === "edit" ? mode.song : undefined}
           submitLabel={mode.kind === "edit" ? "Save changes" : "Create song"}
           busy={busy}
           onCancel={() => setMode({ kind: "list" })}
@@ -372,110 +171,13 @@ function SongsConsole({ pw, onAuthFail, onSignOut }: SongsConsoleProps) {
         />
       ) : null}
 
-      <div className={styles.tableWrap} aria-busy={loading}>
-        {/* Stale-while-revalidate: on a filter/page change we keep the current
-            rows on screen (dimmed) while refetching, instead of blanking the
-            whole table back to skeletons. Skeletons show only on the very first
-            load, when there are no rows to keep. */}
-        <table
-          className={`${styles.table} ${loading && songs.length > 0 ? styles.tableStale : ""}`}
-        >
-          <thead>
-            <tr>
-              <th>Title</th>
-              <th>Artist</th>
-              <th>YouTube ID</th>
-              <th>Start</th>
-              <th>Year</th>
-              <th>Genres</th>
-              <th aria-label="Actions" />
-            </tr>
-          </thead>
-          <tbody>
-            {loading && songs.length === 0 ? (
-              Array.from({ length: 5 }).map((_, i) => (
-                <tr key={`s-${i}`}>
-                  <td>
-                    <Skeleton width="80%" height={16} />
-                  </td>
-                  <td>
-                    <Skeleton width="60%" height={16} />
-                  </td>
-                  <td>
-                    <Skeleton width="100px" height={16} />
-                  </td>
-                  <td>
-                    <Skeleton width="40px" height={16} />
-                  </td>
-                  <td>
-                    <Skeleton width="40px" height={16} />
-                  </td>
-                  <td>
-                    <Skeleton width="120px" height={16} />
-                  </td>
-                  <td />
-                </tr>
-              ))
-            ) : songs.length === 0 ? (
-              <tr>
-                <td colSpan={7} className={styles.empty}>
-                  No songs found.
-                </td>
-              </tr>
-            ) : (
-              songs.map((s) => {
-                const genreNames = (s.genres ?? []).map((g) => g.name);
-                return (
-                  <tr key={s.id}>
-                    <td>{s.title}</td>
-                    <td>{s.artist}</td>
-                    <td className={styles.ytId}>{s.youtube_id}</td>
-                    <td className={styles.startTime}>
-                      {s.start_time > 0 ? `${s.start_time}s` : "—"}
-                    </td>
-                    <td className={styles.startTime}>
-                      {s.release_year != null ? s.release_year : "—"}
-                    </td>
-                    <td>
-                      <div className={styles.genreList}>
-                        {genreNames.length === 0 ? (
-                          <span className={styles.startTime}>—</span>
-                        ) : (
-                          genreNames.map((name) => (
-                            <span key={name} className={styles.genreTag}>
-                              {name}
-                            </span>
-                          ))
-                        )}
-                      </div>
-                    </td>
-                    <td>
-                      <div className={styles.rowActions}>
-                        <button
-                          type="button"
-                          className="btn btn-ghost"
-                          onClick={() => songToFilterEdit(s)}
-                          disabled={busy}
-                        >
-                          Edit
-                        </button>
-                        <button
-                          type="button"
-                          className="btn btn-danger"
-                          onClick={() => setDeleteId(s.id)}
-                          disabled={busy}
-                        >
-                          Delete
-                        </button>
-                      </div>
-                    </td>
-                  </tr>
-                );
-              })
-            )}
-          </tbody>
-        </table>
-      </div>
+      <SongTable
+        songs={songs}
+        loading={loading}
+        busy={busy}
+        onEdit={startEdit}
+        onDelete={(id) => setDeleteId(id)}
+      />
 
       {total > 0 ? (
         <div className={styles.pagination}>
@@ -513,127 +215,5 @@ function SongsConsole({ pw, onAuthFail, onSignOut }: SongsConsoleProps) {
         onCancel={() => setDeleteId(null)}
       />
     </main>
-  );
-}
-
-interface SongFormProps {
-  genres: Genre[];
-  initial: FormState;
-  submitLabel: string;
-  busy: boolean;
-  onCancel: () => void;
-  onSubmit: (payload: SongWritePayload) => void;
-}
-
-function SongForm({ genres, initial, submitLabel, busy, onCancel, onSubmit }: SongFormProps) {
-  const [form, setForm] = useState<FormState>(initial);
-
-  const valid = useMemo(() => formIsValid(form), [form]);
-
-  function toggleGenre(id: string) {
-    setForm((prev) => {
-      const next = new Set(prev.genre_ids);
-      if (next.has(id)) next.delete(id);
-      else next.add(id);
-      return { ...prev, genre_ids: next };
-    });
-  }
-
-  function handleSubmit(e: FormEvent) {
-    e.preventDefault();
-    if (!valid || busy) return;
-    onSubmit(formToPayload(form));
-  }
-
-  return (
-    <form className={styles.formPanel} onSubmit={handleSubmit}>
-      <h2>{submitLabel}</h2>
-      <p className="muted">
-        Tagging a song with the Soundtracks or Israeli Soundtracks genre makes it a +15 “name the
-        film/show” round. For those, put the film/show name in <strong>Artist</strong> (that’s the
-        answer revealed on screen); <strong>Title</strong> is the song/clip name, shown only as a
-        hint. Set Title equal to the film name when there’s no distinct song.
-      </p>
-      <div className={styles.formGrid}>
-        <label className={styles.field}>
-          <span>Title</span>
-          <input
-            type="text"
-            value={form.title}
-            onChange={(e) => setForm({ ...form, title: e.target.value })}
-            maxLength={200}
-            aria-label="Title"
-          />
-        </label>
-        <label className={styles.field}>
-          <span>Artist</span>
-          <input
-            type="text"
-            value={form.artist}
-            onChange={(e) => setForm({ ...form, artist: e.target.value })}
-            maxLength={200}
-            aria-label="Artist"
-          />
-        </label>
-        <label className={styles.field}>
-          <span>YouTube ID (11 chars)</span>
-          <input
-            type="text"
-            value={form.youtube_id}
-            onChange={(e) => setForm({ ...form, youtube_id: e.target.value })}
-            maxLength={11}
-            spellCheck={false}
-            aria-label="YouTube ID"
-          />
-        </label>
-        <label className={styles.field}>
-          <span>Start time (seconds)</span>
-          <input
-            type="number"
-            min={0}
-            value={form.start_time}
-            onChange={(e) => setForm({ ...form, start_time: e.target.value })}
-            aria-label="Start time"
-          />
-        </label>
-        <label className={styles.field}>
-          <span>Release year (optional)</span>
-          <input
-            type="number"
-            min={1900}
-            max={2100}
-            placeholder="e.g. 1985"
-            value={form.release_year}
-            onChange={(e) => setForm({ ...form, release_year: e.target.value })}
-            aria-label="Release year"
-          />
-        </label>
-        <div className={`${styles.field} ${styles.fieldFull}`}>
-          <span>Genres ({form.genre_ids.size} selected)</span>
-          <div className={styles.genres}>
-            {genres.map((g) => {
-              const sel = form.genre_ids.has(g.id);
-              return (
-                <label
-                  key={g.id}
-                  className={`${styles.genreChip} ${sel ? styles.genreSelected : ""}`}
-                >
-                  <input type="checkbox" checked={sel} onChange={() => toggleGenre(g.id)} />
-                  {g.name}
-                </label>
-              );
-            })}
-          </div>
-        </div>
-      </div>
-      <div className={styles.formActions}>
-        <button type="button" className="btn btn-ghost" onClick={onCancel} disabled={busy}>
-          Cancel
-        </button>
-        <button type="submit" className="btn btn-primary" disabled={!valid || busy}>
-          {busy ? "Saving…" : submitLabel}
-        </button>
-      </div>
-    </form>
   );
 }
